Create README.md
Using the ConstraintLayout for the layout and two event listeners for the app logic. After you enter a base amount and a tip percentage, the app will calculate the tip and total for you. We also implement a tip percentage animation and a footer to make our app more unique.

 - Instructure - Rahul Pandey - @rpandey1234 on GITHUB - https://www.youtube.com/watch?v=FjrKMcnKahY